# internal/server/openapi.yaml
openapi: 3.1.0
info:
  title: WorkTime Sync API
  version: 0.1.0
  description: |
    Веб-платформа для актуализации рабочего времени сотрудников.
    Endpoints разделены на публичные (`/auth/*`, `/webhooks/*`) и защищённые JWT (всё остальное).
  contact:
    name: WorkTime Sync team
servers:
  - url: http://localhost:8080
    description: Local dev
...
paths:
...
  /api/v1/teams:
    get:
      tags: [teams]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: ok }
  /api/v1/teams/{id}/availability:
    get:
      tags: [teams]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: tz
          schema: { type: string, example: Europe/Moscow }
      responses:
        '200':
          description: 5 дней × 11 часов × участники
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamAvailability'
  /api/v1/teams/{id}/find-window:
    post:
      tags: [scheduler]
      security: [{ bearerAuth: [] }]
      summary: Подобрать топ-N слотов для встречи команды
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                duration_min: { type: integer, default: 60 }
                days: { type: integer, default: 7 }
                tz: { type: string, example: Europe/Moscow }
                top_n: { type: integer, default: 3 }
      responses:
        '200': { description: ok }
...
  /api/v1/webhooks/{provider}:
    post:
      tags: [webhooks]
      summary: Приём webhook'ов от внешних провайдеров
      description: |
        Без авторизации — провайдеры подписываются на сервере.
        Валидация подписи происходит внутри сервиса.
        MS Graph: validation handshake через ?validationToken=.
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [ical, caldav, google_calendar, ms365, jira, yandex_tracker]
      responses:
        '202': { description: queued }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TeamAvailability:
      type: object
      properties:
        team_id: { type: string, format: uuid }
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        hours: { type: array, items: { type: integer } }
        days: { type: array, items: { type: string } }
        rows:
          type: array
          items:
            type: object
            properties:
              employee_id: { type: string, format: uuid }
              full_name: { type: string }
              timezone: { type: string }
              cells:
                type: array
                items:
                  type: string
                  enum: [free, busy, conflict, off]